# docs/openapi.yaml
openapi: 3.1.0
info:
  title: VOTR API
  version: "1.0.0"
  description: |
    Every position, traceable to the receipt.

    ## The one property everything else rests on

    A quote is never stored as text. Insights store character offsets into
    `documents.raw_text`, and every `quote` in this document is produced by
    slicing the stored document at those offsets at read time. That is why
    `span` is on every card: it is the receipt, not decoration. An insight
    whose quote could not be located in its source was never written.

    ## Versioning, and what "additive only" commits to

    Everything lives under `/api/v1`. The unversioned paths (`/api/feed`,
    `/api/topics`, `/api/profile`, `/api/saved`, `/api/insights/{id}/source`,
    `/api/insights/{id}/share`, `/api/share/{id}`) still work and are the
    same handler object, so they cannot drift apart from the versioned ones.
    They are not documented separately below; they are frozen aliases.

    Within v1:

    - A new field may appear on any response at any time. Ignore what you do
      not recognise.
    - An existing field will not change type, change meaning, or disappear.
    - A new value may be added to any enum here. **Branch with a default
      case**, never exhaustively.
    - A new query parameter may appear. Omitting it keeps the old behaviour.

    Anything that cannot be done additively gets `/api/v2`, and v1 keeps
    answering until the frontend has moved.

    ## Identifying a reader

...

...

...

...
servers:
...
paths:
...
  /api/v1/saved:
    get:
      tags: [reader]
      operationId: listSaved
      summary: The reader's saved cards, newest save first
      parameters:
        - $ref: "#/components/parameters/SessionHeader"
        - name: cursor
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
      responses:
        "200":
          description: |
            The same card shape the feed returns, plus `savedAt` and a `total`.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/FeedPage"
                  - type: object
                    properties:
                      total: { type: integer }
        "400":
          $ref: "#/components/responses/MissingUser"
        "500":
          $ref: "#/components/responses/ServerError"
    post:
      tags: [reader]
      operationId: saveInsight
      summary: Save a card
      description: |
        Idempotent by design, because the client is a swipe on a phone: a
        double-tap, a retry after a dropped connection, and a second tab must
        all converge on one row. A repeat save is 200, a first save is 201, and
        a client retrying after a timeout does not have to tell them apart.
      parameters:
        - $ref: "#/components/parameters/SessionHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [insightId]
              properties:
                insightId: { type: string, format: uuid }
      responses:
        "200":
          description: Already saved.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SaveResult" }
        "201":
          description: Newly saved.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SaveResult" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
    delete:
      tags: [reader]
      operationId: unsaveInsight
      summary: Unsave a card
      description: |
        Removing a save that is not there is a 200, not a 404: the caller
        asked for this card not to be saved, and after this request it is not
        saved.
      parameters:
        - $ref: "#/components/parameters/SessionHeader"
        - name: insightId
          in: query
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Whether a row was removed.
          content:
            application/json:
              schema:
                type: object
                properties:
                  removed: { type: boolean }
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /api/v1/dev/seed:
    get:
      tags: [fixtures]
      operationId: getDemoDataStatus
      summary: What the fixture set contains and whether it is installed
      description: |
        Writes nothing. Returns 404 in production unless `ALLOW_DEMO_SEED` is
        `true`, which is the honest answer: the endpoint does not exist there.
      responses:
        "200":
          description: Fixture status.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DemoStatus" }
        "404":
          description: Seeding is disabled in this environment.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "500":
          $ref: "#/components/responses/ServerError"
    post:
      tags: [fixtures]
      operationId: seedDemoData
      summary: Install the demo fixtures
      description: |
        Idempotent. Documents are keyed by a fixed URL and insights by the
        existing (document, span, card type) dedup index, so calling it twice
        converges rather than duplicating.

        `reset: true` deletes the fixture rows first. It removes only rows this
        fixture set created and never truncates a table, because this has to be
        safe to point at the shared database the rest of the team's real data
        lives in.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reset: { type: boolean, default: false }
      responses:
        "200":
          description: What was installed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SeedResult" }
        "404":
          description: Seeding is disabled in this environment.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "500":
          $ref: "#/components/responses/ServerError"
    delete:
      tags: [fixtures]
      operationId: clearDemoData
      summary: Remove the fixture rows and install nothing
      responses:
        "200":
          description: What was removed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SeedResult" }
        "404":
          description: Seeding is disabled in this environment.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "500":
          $ref: "#/components/responses/ServerError"
components:
  parameters:
    SessionHeader:
      name: x-votr-session
      in: header
      required: false
      schema: { type: string, minLength: 8, maxLength: 128, pattern: "^[A-Za-z0-9_-]+$" }
      description: |
        The anonymous reader id. Also accepted as the `votr_session` cookie or
        a `userId` query parameter, in that order of precedence.
...
  responses:
    BadRequest:
      description: The request is malformed.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    MissingUser:
      description: |
        No reader id. Send an `x-votr-session` header, a `votr_session`
        cookie, or a `userId` query parameter.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: No such thing, or it is not published.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ServerError:
      description: The server failed. Retrying is reasonable.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
          description: For a human. The wording may change; the code will not.
        code:
          type: string
          enum: [bad_request, missing_user, not_found, unprocessable, server_error]
          description: For a switch statement.
        details:
          type: object
          additionalProperties: true
          description: Field-level detail, when the failure is about one input.

    # ---------------------------------------------------------------- enums
    CardType:
      type: string
      enum: [stance, stance_change, factual_claim, voter_relevance]
      description: |
        The feed discriminator, and what `payload` narrows on.

        - `stance` - a position a named person holds.
        - `stance_change` - a position that differs from one already on
          record. Only produced when prior verified stances were given to the
          extractor, so the model cannot decide on its own to allege a change.
        - `factual_claim` - a checkable assertion about the world.
        - `voter_relevance` - a concrete risk or opportunity for a named group.
    InsightFlag:
      type: string
      enum: [NEW, FLIP_FLOP, UNVERIFIED_CLAIM]
      description: |
        The "what changed" badges. A card can carry several: a first-ever
        stance change on a topic is both `NEW` and `FLIP_FLOP`.

        - `NEW` - the first thing on record for this speaker on this topic.
          Assigned at write time and never recomputed, so a card that was the
          first of its kind keeps the badge once later ones arrive. That is
          what a reader scrolling a feed means by "new". Requires a named
          speaker; unattributed cards never carry it.
        - `FLIP_FLOP` - a recorded change of position that cleared the
          confidence bar. Three things must all hold: the pipeline produced a
          `stance_change` card, the two positions are substantively different
          rather than reworded, and the extractor's confidence that a change
          occurred cleared the threshold. A `stance_change` card WITHOUT this
          badge is a shift this app is not willing to call a flip-flop; do not
          label it as one.
        - `UNVERIFIED_CLAIM` - a factual assertion a reader should not take on
          trust: either it is not the kind of thing that can be settled
          quickly, or nothing has settled it yet. Note the gap: a claim checked
          and found FALSE loses this badge, because it is no longer unverified.
          Read `factCheckStatus` for that.
    FactCheckStatus:
      type: string
      enum: [unresolved, supported, disputed, "false"]
      description: |
        Where a factual claim stands against an external checker. Only
        meaningful on `factual_claim` cards; other card types carry
        `unresolved` and nothing should read it.

        - `unresolved` - nothing has checked it. The honest default, and today
          the common case.
        - `supported` - a checker found it holds up.
        - `disputed` - checkers disagree, or it holds up only in part.
        - `false` - a checker found it does not hold up.

        `unresolved` and `supported` are not interchangeable. A claim nobody
        has checked is not a claim that checked out, and collapsing the two is
        the error this app exists to avoid.
    Checkability:
      type: string
      enum: [easily_checkable, requires_expertise, unverifiable]
      description: |
        On `factual_claim` payloads: how settleable the assertion is.

        - `easily_checkable` - a named source could settle it in minutes, such
          as a public statistic or a vote record.
        - `requires_expertise` - settling it needs domain analysis or modelling.
        - `unverifiable` - a prediction, a statement about intent, or a value
          judgment. Not a failure to check it; it is not the kind of thing that
          can be checked, and saying so is useful.
    ConfidenceLabel:
      type: string
      enum: [high, medium, low]
      description: |
        The bucketed `claimSupportConfidence`: how directly the quote supports
        this card's claim. Default cut points are `high >= 0.8`,
        `medium >= 0.5`, `low` below, and they are configurable server-side.

        `low` is the one that changes how a card reads, and it is also what an
        UNRATED card gets - an insight nothing rated is exactly the case the
        nudge exists for. Prefer `presentationMode` over branching on this.
    PresentationMode:
      type: string
      enum: [stated, nudge_verify]
      description: |
        What a client may do with the card. Derived from `confidenceLabel` on
        the server so a threshold change ships without a client release.

        - `stated` - render the claim as the card states it.
        - `nudge_verify` - **never present it as flatly stated fact.** Lead
          with the quote, show the check-it-yourself affordance, and link the
          source. This is the contract the API is making on the reader's
          behalf, not a styling hint.
    QuoteVerification:
      type: string
      enum: [exact, normalized, transcript, fuzzy]
      description: |
        Which rung of the verification ladder located this quote in its
        document. There is deliberately no failure value: a quote that could
        not be located was never stored as an insight.

        `fuzzy` means the model's string had to be corrected to match the
        document, and `quoteSimilarity` says by how much.
    Attribution:
      type: string
      enum: [own_words, characterization]
      description: |
        `own_words` when the quote is the person speaking or writing;
        `characterization` when a reporter or third party is describing them.
        Worth surfacing: "reported to support X" is a weaker thing than "said
        X", and the card should not flatten the difference.

    # ---------------------------------------------------------------- feed
    FeedPage:
      type: object
      required: [items, nextCursor, sort]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/FeedItem" }
        nextCursor:
          oneOf: [{ type: string }, { type: "null" }]
          description: Pass back as `cursor`. Null means the run is over.
        sort: { type: string }
        personalization:
          type: object
          properties:
            userId:
              oneOf: [{ type: string }, { type: "null" }]
            selectedTopics:
              type: array
              items: { type: string }
            topicMode: { type: string, enum: [boost, strict] }
            applied:
              type: boolean
              description: False means no profile, so nothing was re-ranked.
    FeedItem:
      type: object
      required:
        [id, cardType, payload, headline, plainLanguage, quote, span,
         quoteVerified, issueTag, topics, attribution, flags, factCheckStatus,
         confidenceLabel, verifyYourself, presentationMode, sourceDiversity,
         documentId, sourceName, sourceUrl, createdAt]
      properties:
        id: { type: string, format: uuid }
        cardType: { $ref: "#/components/schemas/CardType" }
        payload:
          description: Kind-specific fields. Narrow on `cardType`.
          oneOf:
            - $ref: "#/components/schemas/StancePayload"
            - $ref: "#/components/schemas/StanceChangePayload"
            - $ref: "#/components/schemas/FactualClaimPayload"
            - $ref: "#/components/schemas/VoterRelevancePayload"
          discriminator:
            propertyName: cardType
        headline:
          type: string
          description: The card's headline claim, whatever its kind.
        positionText:
          type: string
          description: Alias of `headline`, for the name the first UI used.
        plainLanguage:
          type: string
          description: One neutral sentence at a low reading level.
        plainLanguageSummary:
          oneOf: [{ type: string }, { type: "null" }]
          description: |
            The rewrite pass's sentence, or null when it did not run or did not
            produce something usable. Never a silent copy of `plainLanguage`: a
            client has to be able to tell "not rewritten" from "rewritten".
        readingLevelEstimate:
          oneOf: [{ type: number }, { type: "null" }]
        quote:
          type: string
          description: |
            Sliced out of the stored document at `span` at read time. There is
            no stored quote string anywhere in this system to drift.
        span:
          type: object
          required: [start, end]
          description: Character offsets into the stored document. The receipt.
          properties:
            start: { type: integer }
            end: { type: integer }
        quoteVerified: { $ref: "#/components/schemas/QuoteVerification" }
        quoteSimilarity:
          oneOf: [{ type: number }, { type: "null" }]
          description: 1 on the lossless rungs, the measured ratio on `fuzzy`.
        issueTag:
          type: string
          description: The primary issue slug. Also present in `topics`.
        topics:
          type: array
          items: { type: string }
          description: Every issue this insight is tagged with, primary first.
        attribution: { $ref: "#/components/schemas/Attribution" }
        flag:
          description: |
            Legacy mirror of `flags[0]`. New code reads `flags`.
          oneOf:
            - $ref: "#/components/schemas/InsightFlag"
            - type: "null"
        flags:
          type: array
          items: { $ref: "#/components/schemas/InsightFlag" }
        factCheckStatus: { $ref: "#/components/schemas/FactCheckStatus" }
        claimSupportConfidence:
          oneOf: [{ type: number, minimum: 0, maximum: 1 }, { type: "null" }]
          description: |
            How directly the quote supports this card's claim. Null on a row
            written before the extractor was asked. **Branch on
            `presentationMode`, not on this.**
        confidenceLabel: { $ref: "#/components/schemas/ConfidenceLabel" }
        verifyYourself:
          type: boolean
          description: True exactly when `confidenceLabel` is `low`.
        presentationMode: { $ref: "#/components/schemas/PresentationMode" }
        sourceDiversity:
          type: object
          required: [count, sources]
          description: |
            How many independent outlets report this insight.
          properties:
            count:
              type: integer
              minimum: 1
              description: |
                Distinct sources INCLUDING this card's own, so 1 means "one
                outlet, nothing has confirmed it" and there is no 0. Counted
                by source, not by document: two articles from one outlet are
                one. `count` can exceed `sources.length`, which is capped by
                `sourceDiversityLimit`; render "+N more" from the count.
            sources:
              type: array
              description: The OTHER outlets, capped.
              items:
                type: object
                properties:
                  sourceName: { type: string }
                  sourceUrl: { type: string, format: uri }
        matchesInterests:
          type: boolean
          description: |
            Whether this card matched the reader's selected topics. Returned so
            "because you follow housing" can be shown honestly rather than
            guessed at.
        judgeRating:
          oneOf: [{ type: integer }, { type: "null" }]
        relevanceScore: { type: number }
        candidateName:
          oneOf: [{ type: string }, { type: "null" }]
          description: Null when the document names no one, as claims often do.
        speakerId:
          oneOf: [{ type: string, format: uuid }, { type: "null" }]
          description: |
            The person, not the ballot line. This is the id the timeline
            endpoint takes. Null on an unattributed card.
        documentId: { type: string, format: uuid }
        documentTitle: { type: string }
        sourceName: { type: string }
        sourceUrl: { type: string, format: uri }
        publishedAt:
          oneOf: [{ type: string, format: date-time }, { type: "null" }]
        createdAt: { type: string, format: date-time }
        isSynthetic:
          type: boolean
          description: |
            True for demo and fixture documents. Badge it: a synthetic source
            must never be mistaken for real coverage.
        alsoCited:
          type: array
          description: |
            Other insights citing this exact quote, collapsed into this card.
            Non-empty means the same words support more than one kind of
            insight and the feed is showing the most informative one.
          items:
            type: object
            properties:
              id: { type: string, format: uuid }
              cardType: { $ref: "#/components/schemas/CardType" }
        saved:
          type: boolean
          description: Present only when the request carried a reader id.
        savedAt:
          oneOf: [{ type: string, format: date-time }, { type: "null" }]
          description: Set only on a saved-list query.
    StancePayload:
      type: object
      required: [cardType]
      properties:
        cardType: { type: string, const: stance }
    StanceChangePayload:
      type: object
      required: [cardType, previousPosition]
      properties:
        cardType: { type: string, const: stance_change }
        previousPosition:
          type: string
          description: |
            Taken from the prior verified record the extractor was shown, never
            from the model's own knowledge.
        priorInsightId:
          oneOf: [{ type: string, format: uuid }, { type: "null" }]
          description: |
            The prior insight this was measured against, when the model named
            one that resolved to a real row. Null otherwise: the card still
            stands on its own verified quote, it just loses that edge.
    FactualClaimPayload:
      type: object
      required: [cardType, checkability]
      properties:
        cardType: { type: string, const: factual_claim }
        checkability: { $ref: "#/components/schemas/Checkability" }
    VoterRelevancePayload:
      type: object
      required: [cardType, relevanceKind, affectedGroup]
      properties:
        cardType: { type: string, const: voter_relevance }
        relevanceKind: { type: string, enum: [risk, opportunity] }
        affectedGroup:
          type: string
          description: As specific as the document supports. Never "everyone".
...
    SaveResult:
      type: object
      properties:
        insightId: { type: string, format: uuid }
        savedAt: { type: string, format: date-time }
        alreadySaved:
          type: boolean
          description: True on a repeat save, which is a 200 rather than a 201.
...
    DemoStatus:
      type: object
      properties:
        installed: { type: boolean }
        documents: { type: integer }
        insights: { type: integer }
        speakers: { type: integer }
        expected:
          type: object
          properties:
            documents: { type: integer }
            insights: { type: integer }
            speakers: { type: integer }
        urlPrefix:
          type: string
          description: |
            Every fixture document's URL starts with this, which is how a reset
            finds exactly the rows it created and nothing else.
    SeedResult:
      type: object
      properties:
        speakers: { type: integer }
        documents: { type: integer }
        insights: { type: integer }
        topicLinks: { type: integer }
        corroborationLinks: { type: integer }
        corroborated:
          type: integer
          description: Insights confirmed by more than one source after seeding.
        reset: { type: boolean }
        deletedDocuments: { type: integer }
        deletedSpeakers: { type: integer }